# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Product Managment (Modular API)",
        "description": "API documentation for best practice modular project",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Main server API"
        }
    ],
    "paths": {
...
        "/api/v1/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get list of all products",
                "operationId": "4998de8c2880988fa95bb7c671fe5ebb",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/ProductResource"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Create a new product",
                "operationId": "da23ef06d62bf6f0498b8ac9311f9633",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Product created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/v1/products/{id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Get a specific product by ID",
                "operationId": "fb22ebe7cac946ff7e724fb0abac56b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Update an existing product",
                "operationId": "720977ea02b13031fe7225d565ce7a14",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateProductRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Product updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ProductResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete a product",
                "operationId": "ccf3b954a27a07214f085803ed45f79f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Product ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Product deleted successfully"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreProductRequest": {
                "title": "Store Product Request",
                "description": "Request schema for storing a new product",
                "required": [
                    "name",
                    "price"
                ],
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "laptop"
                    },
                    "description": {
                        "type": "string",
                        "example": "A high-quality laptop",
                        "nullable": true
                    },
                    "price": {
                        "type": "integer",
                        "example": 1500000
                    }
                },
                "type": "object"
            },
            "UpdateProductRequest": {
                "title": "Update Product Request",
                "description": "Request schema for updating an existing product",
                "properties": {
                    "name": {
                        "type": "string",
                        "example": "laptop"
                    },
                    "description": {
                        "type": "string",
                        "example": "An updated high-quality laptop",
                        "nullable": true
                    },
                    "price": {
                        "type": "integer",
                        "example": 1400000
                    }
                },
                "type": "object"
            },
            "ProductResource": {
                "title": "Product Resource",
                "description": "Product resource representation",
                "properties": {
                    "data": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "name": {
                                "type": "string",
                                "example": "laptop"
                            },
                            "description": {
                                "type": "string",
                                "example": "A high-quality laptop",
                                "nullable": true
                            },
                            "price": {
                                "type": "integer",
                                "example": 1500000
                            },
                            "created_at": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2025-10-20 14:30:00"
                            },
                            "updated_at": {
                                "type": "string",
                                "format": "date-time",
                                "example": "2025-10-20 14:30:00"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Import your JWT token in the format 'Bearer'",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}